requirements.txt
tensorflow
keras
tensorflow-datasets
numpy
pandas
scikit-learn
Flask
Pillow
matplotlib
opencv-python
